github: Skip caching in setup-msys2
The stored caches are quite large (around 200 MB each), and they
don't speed up the setup at all (on the contrary, fetching and
storing the cache actually can make things marginally slower).

Using a cache can reduce the load on the upstream package repository
though.

By not using a cache for setup-msys2, we keep more cache space
available for the actual LLVM builds, which take a couple orders
of magnitude longer to run.